# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Budgetarian API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorEnvelope {
  error: string;
}

export interface Expense {
  id: number;
  userId: string;
  amount: number;
  category: string;
  description: string;
  date: string;
  notes?: string | null;
  createdAt: string;
}

export interface CategoryTotal {
  category: string;
  total: number;
}

export interface DashboardSummary {
  /** Total expense amount this month */
  totalExpensesThisMonth: number;
  /** Total amount saved across all active goals */
  totalSaved: number;
  /** Number of active savings goals */
  activeGoals: number;
  /** Number of installments due within 7 days */
  upcomingDues: number;
  /** Number of warranties expiring within 30 days */
  expiringSoonCount: number;
  recentExpenses?: Expense[];
  categoryBreakdown?: CategoryTotal[];
}

export interface ListExpensesResponse {
  expenses: Expense[];
  total: number;
}

export interface CreateExpenseBody {
  /** @minimum 0.01 */
  amount: number;
  /** @minLength 1 */
  category: string;
  /** @minLength 1 */
  description: string;
  date: string;
  notes?: string;
}

export interface UpdateExpenseBody {
  /** @minimum 0.01 */
  amount?: number;
  category?: string;
  description?: string;
  date?: string;
  notes?: string;
}

export type ExpenseSummaryMonthlyTrendItem = {
  month: string;
  total: number;
};

export interface ExpenseSummary {
  thisMonth: number;
  lastMonth: number;
  byCategory: CategoryTotal[];
  monthlyTrend: ExpenseSummaryMonthlyTrendItem[];
}

export type GoalPeriod = (typeof GoalPeriod)[keyof typeof GoalPeriod];

export const GoalPeriod = {
  weekly: "weekly",
  monthly: "monthly",
} as const;

export interface Goal {
  id: number;
  userId: string;
  name: string;
  targetAmount: number;
  savedAmount: number;
  period: GoalPeriod;
  notes?: string | null;
  createdAt: string;
}

export interface ListGoalsResponse {
  goals: Goal[];
}

export type CreateGoalBodyPeriod =
  (typeof CreateGoalBodyPeriod)[keyof typeof CreateGoalBodyPeriod];

export const CreateGoalBodyPeriod = {
  weekly: "weekly",
  monthly: "monthly",
} as const;

export interface CreateGoalBody {
  /** @minLength 1 */
  name: string;
  /** @minimum 0.01 */
  targetAmount: number;
  /** @minimum 0 */
  savedAmount?: number;
  period: CreateGoalBodyPeriod;
  notes?: string;
}

export type UpdateGoalBodyPeriod =
  (typeof UpdateGoalBodyPeriod)[keyof typeof UpdateGoalBodyPeriod];

export const UpdateGoalBodyPeriod = {
  weekly: "weekly",
  monthly: "monthly",
} as const;

export interface UpdateGoalBody {
  name?: string;
  /** @minimum 0.01 */
  targetAmount?: number;
  /** @minimum 0 */
  savedAmount?: number;
  period?: UpdateGoalBodyPeriod;
  notes?: string;
}

export type InstallmentStatus =
  (typeof InstallmentStatus)[keyof typeof InstallmentStatus];

export const InstallmentStatus = {
  pending: "pending",
  paid: "paid",
  overdue: "overdue",
} as const;

export interface Installment {
  id: number;
  userId: string;
  name: string;
  amount: number;
  dueDate: string;
  status: InstallmentStatus;
  notes?: string | null;
  createdAt: string;
}

export interface ListInstallmentsResponse {
  installments: Installment[];
}

export type CreateInstallmentBodyStatus =
  (typeof CreateInstallmentBodyStatus)[keyof typeof CreateInstallmentBodyStatus];

export const CreateInstallmentBodyStatus = {
  pending: "pending",
  paid: "paid",
  overdue: "overdue",
} as const;

export interface CreateInstallmentBody {
  /** @minLength 1 */
  name: string;
  /** @minimum 0.01 */
  amount: number;
  dueDate: string;
  status?: CreateInstallmentBodyStatus;
  notes?: string;
}

export type UpdateInstallmentBodyStatus =
  (typeof UpdateInstallmentBodyStatus)[keyof typeof UpdateInstallmentBodyStatus];

export const UpdateInstallmentBodyStatus = {
  pending: "pending",
  paid: "paid",
  overdue: "overdue",
} as const;

export interface UpdateInstallmentBody {
  name?: string;
  /** @minimum 0.01 */
  amount?: number;
  dueDate?: string;
  status?: UpdateInstallmentBodyStatus;
  notes?: string;
}

export interface Warranty {
  id: number;
  userId: string;
  productName: string;
  store?: string | null;
  purchaseDate: string;
  expiryDate: string;
  /** Object path for the receipt image stored in object storage */
  receiptPath?: string | null;
  notes?: string | null;
  createdAt: string;
}

export interface ListWarrantiesResponse {
  warranties: Warranty[];
}

export interface CreateWarrantyBody {
  /** @minLength 1 */
  productName: string;
  store?: string;
  purchaseDate: string;
  expiryDate: string;
  /** Object path returned from object storage upload */
  receiptPath?: string;
  notes?: string;
}

export interface UpdateWarrantyBody {
  productName?: string;
  store?: string;
  purchaseDate?: string;
  expiryDate?: string;
  receiptPath?: string;
  notes?: string;
}

export type TipCategory = (typeof TipCategory)[keyof typeof TipCategory];

export const TipCategory = {
  saving: "saving",
  income: "income",
  budgeting: "budgeting",
  investing: "investing",
} as const;

export interface Tip {
  id: number;
  title: string;
  description: string;
  category: TipCategory;
}

export interface ListTipsResponse {
  tips: Tip[];
}

export interface UploadUrlRequest {
  /** @minLength 1 */
  name: string;
  /** @minimum 1 */
  size: number;
  /** @minLength 1 */
  contentType: string;
}

export interface UploadUrlResponse {
  uploadURL: string;
  objectPath: string;
  metadata?: UploadUrlRequest;
}

export type ListExpensesParams = {
  limit?: number;
  offset?: number;
  category?: string;
};