# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Budgetarian API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /expenses:
    get:
      operationId: listExpenses
      tags: [expenses]
      summary: List expenses
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of expenses
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListExpensesResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    post:
      operationId: createExpense
      tags: [expenses]
      summary: Create an expense
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateExpenseBody"
      responses:
        "201":
          description: Expense created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /expenses/summary:
    get:
      operationId: getExpenseSummary
      tags: [expenses]
      summary: Get expense summary by category and time period
      responses:
        "200":
          description: Expense summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExpenseSummary"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /expenses/{id}:
    get:
      operationId: getExpense
      tags: [expenses]
      summary: Get an expense by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Expense
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    patch:
      operationId: updateExpense
      tags: [expenses]
      summary: Update an expense
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateExpenseBody"
      responses:
        "200":
          description: Updated expense
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    delete:
      operationId: deleteExpense
      tags: [expenses]
      summary: Delete an expense
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  # === SAVINGS GOALS ===
  /goals:
    get:
      operationId: listGoals
      tags: [goals]
      summary: List savings goals
      responses:
        "200":
          description: List of savings goals
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListGoalsResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    post:
      operationId: createGoal
      tags: [goals]
      summary: Create a savings goal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGoalBody"
      responses:
        "201":
          description: Goal created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /goals/{id}:
    patch:
      operationId: updateGoal
      tags: [goals]
      summary: Update a savings goal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateGoalBody"
      responses:
        "200":
          description: Updated goal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    delete:
      operationId: deleteGoal
      tags: [goals]
      summary: Delete a savings goal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  # === INSTALLMENTS ===
  /installments:
    get:
      operationId: listInstallments
      tags: [installments]
      summary: List installment entries
      responses:
        "200":
          description: List of installments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListInstallmentsResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    post:
      operationId: createInstallment
      tags: [installments]
      summary: Create an installment entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInstallmentBody"
      responses:
        "201":
          description: Installment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Installment"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /installments/upcoming:
    get:
      operationId: getUpcomingInstallments
      tags: [installments]
      summary: Get installments due within 7 days
      responses:
        "200":
          description: Upcoming installments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListInstallmentsResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /installments/{id}:
    patch:
      operationId: updateInstallment
      tags: [installments]
      summary: Update an installment entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateInstallmentBody"
      responses:
        "200":
          description: Updated installment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Installment"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    delete:
      operationId: deleteInstallment
      tags: [installments]
      summary: Delete an installment entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  # === WARRANTIES ===
...
components:
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # === DASHBOARD ===
...
    CategoryTotal:
      type: object
      required: [category, total]
      properties:
        category:
          type: string
        total:
          type: number

    # === EXPENSES ===
    Expense:
      type: object
      required: [id, userId, amount, category, description, date, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: string
        amount:
          type: number
        category:
          type: string
        description:
          type: string
        date:
          type: string
          format: date
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
    ListExpensesResponse:
      type: object
      required: [expenses, total]
      properties:
        expenses:
          type: array
          items:
            $ref: "#/components/schemas/Expense"
        total:
          type: integer
    CreateExpenseBody:
      type: object
      required: [amount, category, description, date]
      properties:
        amount:
          type: number
          minimum: 0.01
        category:
          type: string
          minLength: 1
        description:
          type: string
          minLength: 1
        date:
          type: string
          format: date
        notes:
          type: string
    UpdateExpenseBody:
      type: object
      properties:
        amount:
          type: number
          minimum: 0.01
        category:
          type: string
        description:
          type: string
        date:
          type: string
          format: date
        notes:
          type: string
    ExpenseSummary:
      type: object
      required: [thisMonth, lastMonth, byCategory, monthlyTrend]
      properties:
        thisMonth:
          type: number
        lastMonth:
          type: number
        byCategory:
          type: array
          items:
            $ref: "#/components/schemas/CategoryTotal"
        monthlyTrend:
          type: array
          items:
            type: object
            required: [month, total]
            properties:
              month:
                type: string
              total:
                type: number

    # === GOALS ===
    Goal:
      type: object
      required: [id, userId, name, targetAmount, savedAmount, period, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: string
        name:
          type: string
        targetAmount:
          type: number
        savedAmount:
          type: number
        period:
          type: string
          enum: [weekly, monthly]
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
    ListGoalsResponse:
      type: object
      required: [goals]
      properties:
        goals:
          type: array
          items:
            $ref: "#/components/schemas/Goal"
    CreateGoalBody:
      type: object
      required: [name, targetAmount, period]
      properties:
        name:
          type: string
          minLength: 1
        targetAmount:
          type: number
          minimum: 0.01
        savedAmount:
          type: number
          minimum: 0
        period:
          type: string
          enum: [weekly, monthly]
        notes:
          type: string
    UpdateGoalBody:
      type: object
      properties:
        name:
          type: string
        targetAmount:
          type: number
          minimum: 0.01
        savedAmount:
          type: number
          minimum: 0
        period:
          type: string
          enum: [weekly, monthly]
        notes:
          type: string

    # === INSTALLMENTS ===
    Installment:
      type: object
      required: [id, userId, name, amount, dueDate, status, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: string
        name:
          type: string
        amount:
          type: number
        dueDate:
          type: string
          format: date
        status:
          type: string
          enum: [pending, paid, overdue]
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
    ListInstallmentsResponse:
      type: object
      required: [installments]
      properties:
        installments:
          type: array
          items:
            $ref: "#/components/schemas/Installment"
    CreateInstallmentBody:
      type: object
      required: [name, amount, dueDate]
      properties:
        name:
          type: string
          minLength: 1
        amount:
          type: number
          minimum: 0.01
        dueDate:
          type: string
          format: date
        status:
          type: string
          enum: [pending, paid, overdue]
        notes:
          type: string
    UpdateInstallmentBody:
      type: object
      properties:
        name:
          type: string
        amount:
          type: number
          minimum: 0.01
        dueDate:
          type: string
          format: date
        status:
          type: string
          enum: [pending, paid, overdue]
        notes:
          type: string

    # === WARRANTIES ===
...