# lib/api-zod/src/generated/types/createExpenseBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Budgetarian API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateExpenseBody {
  /** @minimum 0.01 */
  amount: number;
  /** @minLength 1 */
  category: string;
  /** @minLength 1 */
  description: string;
  date: Date;
  notes?: string;
}