# lib/api-zod/src/generated/types/createGoalBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Budgetarian API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateGoalBodyPeriod } from "./createGoalBodyPeriod";

export interface CreateGoalBody {
  /** @minLength 1 */
  name: string;
  /** @minimum 0.01 */
  targetAmount: number;
  /** @minimum 0 */
  savedAmount?: number;
  period: CreateGoalBodyPeriod;
  notes?: string;
}