# lib/api-zod/src/generated/types/listGoalsResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Budgetarian API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Goal } from "./goal";

export interface ListGoalsResponse {
  goals: Goal[];
}