# lib/api-zod/src/generated/types/updateGoalBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Budgetarian API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateGoalBodyPeriod } from "./updateGoalBodyPeriod";

export interface UpdateGoalBody {
  name?: string;
  /** @minimum 0.01 */
  targetAmount?: number;
  /** @minimum 0 */
  savedAmount?: number;
  period?: UpdateGoalBodyPeriod;
  notes?: string;
}